requirements.txt
# Core inference runtime
vllm==0.18.0
torch==2.10.0
torchaudio==2.10.0
torchvision==0.25.0
transformers==5.3.0
bitsandbytes==0.48.1
datasets==4.8.4
peft==0.18.1

# Web framework & API
# fastapi==0.135.1
# uvicorn==0.41.0

# AI SDKs for proxy
# anthropic==0.84.0
# openai==2.24.0

# CUDA/NVIDIA stack
# cuda-python==12.9.4
# cupy-cuda12x==14.0.1
# nvidia-cublas-cu12==12.8.4.1
# nvidia-cuda-cupti-cu12==12.8.90
# nvidia-cuda-nvrtc-cu12==12.8.93
# nvidia-cuda-runtime-cu12==12.8.90
# nvidia-cudnn-cu12==9.10.2.21
# nvidia-cudnn-frontend==1.19.1
# nvidia-cufft-cu12==11.3.3.83
# nvidia-cufile-cu12==1.13.1.3
# nvidia-curand-cu12==10.3.9.90
# nvidia-cusolver-cu12==11.7.3.90
# nvidia-cusparse-cu12==12.5.8.93
# nvidia-nccl-cu12==2.27.5
# nvidia-nvjitlink-cu12==12.8.93

# ML frameworks
# triton==3.6.0
# xformers
accelerate==1.13.0
# safetensors==0.7.0
# flashinfer-python==0.6.4  # 与 vllm 0.17.0 + transformers 4.56.0 冲突
huggingface_hub

# # Supporting libraries
numpy>=2.0,<2.3
scipy==1.17.1
pillow==12.1.1
pyyaml==6.0.3
requests==2.32.5
httpx==0.28.1
pydantic==2.12.5
pydantic-settings==2.13.1
jinja2==3.1.6
tiktoken==0.12.0
tokenizers==0.22.2
sentencepiece==0.2.1
tqdm==4.67.3
loguru==0.7.3

# # Infrastructure
ray==2.54.0
psutil==7.2.2
packaging==26.0

# # FastAPI utilities
python-multipart==0.0.22
prometheus-fastapi-instrumentator==7.1.0
sse-starlette==3.3.2
